+++ tests/engines/test_httpx.py
@@ -7,12 +7,12 @@ from httpx import AsyncClient
 from httpx import Client
 from scim2_models import PatchOp
 from scim2_models import PatchOperation
+from scim2_models import SCIMException
 from scim2_models import SearchRequest
 from scim2_models import ServiceProviderConfig
 
 from scim2_client.engines.httpx import AsyncSCIMClient
 from scim2_client.engines.httpx import SyncSCIMClient
-from scim2_client.errors import SCIMResponseErrorObject
 
 scim2_server = pytest.importorskip("scim2_server")
 from scim2_server.backend import InMemoryBackend  # noqa: E402
@@ -93,7 +93,7 @@ def test_sync_engine(server):
     assert queried_user.display_name == "patched name"
 
     scim_client.delete(User, response_user.id)
-    with pytest.raises(SCIMResponseErrorObject):
+    with pytest.raises(SCIMException):
         scim_client.query(User, response_user.id)
 
 
@@ -152,5 +152,5 @@ async def test_async_engine(server):
     assert queried_user.display_name == "async patched name"
 
     await scim_client.delete(User, response_user.id)
-    with pytest.raises(SCIMResponseErrorObject):
+    with pytest.raises(SCIMException):
         await scim_client.query(User, response_user.id)